+++ src/components/Sidebar/SidebarWrapper/SidebarWrapper.stories.tsx
@@ -6,8 +6,7 @@ import { Default as SidebarFooter } from "../SidebarFooter/SidebarFooter.stories
 import { Default as SidebarItem } from "../SidebarItem/SidebarItem.stories";
 import { Default as SidebarItemText } from "../SidebarItemText/SidebarItemText.stories";
 import { Default as SidebarItemIcon } from "../SidebarItemIcon/SidebarItemIcon.stories";
-import { Default as SidebarItemsGroup } from "../SidebarItemsGroup/SidebarItemsGroup.stories";
-import { Default as SidebarItemsGroupHeader } from "../SidebarItemsGroupHeader/SidebarItemsGroupHeader.stories";
+import { Default as ExpandableContent } from "../../ExpandableContent/ExpandableContent.stories";
 
 const Template: FictoanStory<typeof SidebarWrapper> = (args) => <SidebarWrapper {...args} />;
 Template.displayName = SidebarWrapper.displayName;
@@ -30,14 +29,12 @@ Default.args = {
             <SidebarItemText linkText="Item 3"></SidebarItemText>
         </SidebarItem>,
 
-        <SidebarItemsGroup>
-            <SidebarItemsGroupHeader>
-                <SidebarItem>
-                    <SidebarItemIcon></SidebarItemIcon>
-                    <SidebarItemText linkText="Item Summary"></SidebarItemText>
-                </SidebarItem>
-            </SidebarItemsGroupHeader>
-
+        <ExpandableContent summary={(
+            <SidebarItem>
+                <SidebarItemIcon></SidebarItemIcon>
+                <SidebarItemText linkText="Item Summary" weight="600"></SidebarItemText>
+            </SidebarItem>
+        )}>
             <SidebarItem>
                 <SidebarItemIcon></SidebarItemIcon>
                 <SidebarItemText linkText="Nested Item 1"></SidebarItemText>
@@ -46,7 +43,7 @@ Default.args = {
                 <SidebarItemIcon></SidebarItemIcon>
                 <SidebarItemText linkText="Nested Item 2"></SidebarItemText>
             </SidebarItem>
-        </SidebarItemsGroup>,
+        </ExpandableContent>,
 
         <SidebarFooter padding="micro">
             Sidebar footer